bluetooth init

+++ LTS_Nucleo_ReadEPC/Core/Inc/epc901.h
@@ -10,8 +10,7 @@
 
 /* Private function prototypes -----------------------------------------------*/
 void epc901_init();
-void epc901_getData(int shutterTime);
-void
-usDelay(uint16_t delayTime_us);
+void epc901_getData(uint16_t shutterTime);
+void usDelay(uint16_t delayTime_us);
 
 #endif /* CCD_EPC_H_ */
